+++ packages/fedify/src/federation/middleware.test.ts
@@ -61,7 +61,7 @@ test("createFederation()", async (t) => {
     assertThrows(() =>
       createFederation<number>({
         kv,
-        contextLoader: mockDocumentLoader,
+        contextLoaderFactory: () => mockDocumentLoader,
         allowPrivateAddress: true,
       }), TypeError);
     assertThrows(() =>
@@ -545,7 +545,7 @@ test({
         kv,
         origin: "https://ap.example.com",
         documentLoaderFactory: () => mockDocumentLoader,
-        contextLoader: mockDocumentLoader,
+        contextLoaderFactory: () => mockDocumentLoader,
       });
       const ctx = federation.createContext(
         new URL("https://example.com:1234/"),
@@ -1639,7 +1639,7 @@ test("FederationImpl.sendActivity()", async (t) => {
   const kv = new MemoryKvStore();
   const federation = new FederationImpl<void>({
     kv,
-    contextLoader: mockDocumentLoader,
+    contextLoaderFactory: () => mockDocumentLoader,
   });
   const context = federation.createContext(new URL("https://example.com/"));
 
@@ -2009,7 +2009,7 @@ test("ContextImpl.sendActivity()", async (t) => {
   const kv = new MemoryKvStore();
   const federation = new FederationImpl<void>({
     kv,
-    contextLoader: mockDocumentLoader,
+    contextLoaderFactory: () => mockDocumentLoader,
   });
 
   federation
@@ -2168,7 +2168,7 @@ test("ContextImpl.sendActivity()", async (t) => {
   };
   const federation2 = new FederationImpl<void>({
     kv,
-    contextLoader: mockDocumentLoader,
+    contextLoaderFactory: () => mockDocumentLoader,
     queue,
   });
   federation2
@@ -2623,7 +2623,7 @@ test("InboxContextImpl.forwardActivity()", async (t) => {
   const kv = new MemoryKvStore();
   const federation = new FederationImpl<void>({
     kv,
-    contextLoader: mockDocumentLoader,
+    contextLoaderFactory: () => mockDocumentLoader,
   });
 
   await t.step("skip", async () => {